+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.1.2] - 2026-01-28
-- fix negative values getWatt(), getVWave() and getIWave().
+- fix negative values getWatt, getVWave, getIWave.